Refactor: Separate detailed logic from DetectUtility to caller classes
