+++ src/main/java/com/sprint/mission/discodeit/service/file/FileMessageService.java
@@ -18,7 +18,7 @@ public class FileMessageService implements MessageService {
 
     @Override
     public Message createMessage(String contents, User user, Channel channel) {
-        if (!DetectUtility.detectMessageCreate(contents, user, channel)) {
+        if (!detectMessageCreate(contents, user, channel)) {
             throw new RuntimeException("<실패: 메시지 생성에 실패하였습니다.>");
         }
         return fileMessageRepository.create(contents, user, channel);
@@ -42,7 +42,7 @@ public class FileMessageService implements MessageService {
 
     @Override
     public void updateMessageContentsByMessage(Message message, User user, String newContents) {
-        if (!DetectUtility.detectMessageUpdate(message, user, newContents)) {
+        if (!detectMessageUpdate(message, user, newContents)) {
             ErrorMessageUtility.printErrorMessage();
             return;
         }
@@ -52,7 +52,7 @@ public class FileMessageService implements MessageService {
     // Delete
     @Override
     public void deleteMessageByMessage(Message message, User user, Channel channel) {
-        if (!DetectUtility.detectMessageDelete(message, user, channel)) {
+        if (!detectMessageDelete(message, user, channel)) {
             ErrorMessageUtility.printErrorMessage();
             return;
         }
@@ -62,5 +62,33 @@ public class FileMessageService implements MessageService {
     @Override
     public void deleteAllMessages() {
         fileMessageRepository.deleteAll();
+    }
+
+    public static boolean detectMessageCreate(String contents, User user, Channel channel) {
+        boolean detected = DetectUtility.detect(contents, user, channel);
+        boolean matches = user.getChannels().contains(channel) && channel.getUsers().contains(user);
+        return detected && matches;
+    }
+
+    public static boolean detectMessageUpdate(Message message, User user, String newContents) {
+        boolean detected = DetectUtility.detect(message) && DetectUtility.detect(newContents, user);
+        boolean matches = message.getUser().getId().equals(user.getId());
+        if (detected && matches) {
+            return true;
+        } else {
+            ErrorMessageUtility.printErrorMessage();
+            return false;
+        }
+    }
+
+    public static boolean detectMessageDelete(Message message, User user, Channel channel) {
+        boolean detected = DetectUtility.detect(message) && DetectUtility.detect(user) && DetectUtility.detect(channel);
+        boolean matches = message.getUser().getId().equals(user.getId()) && message.getChannel().getId().equals(channel.getId());
+        if (detected && matches) {
+            return true;
+        } else {
+            ErrorMessageUtility.printErrorMessage();
+            return false;
+        }
     }
 }
