+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFUserService.java
@@ -61,7 +61,7 @@ public class JCFUserService implements UserService {
     // 참여중인 채널 관련 메서드
     // 채널 참가
     public void joinChannelOnlyActiveUser(User user, Channel channel) {
-        if (!DetectUtility.detectJoinChannel(user, channel)) {
+        if (!detectJoinChannel(user, channel)) {
             ErrorMessageUtility.printErrorMessage();
             return;
         }
@@ -70,10 +70,22 @@ public class JCFUserService implements UserService {
 
     // 채널 나가기
     public void outChannelOnlyActiveUser(User user, Channel channel) {
-        if (!DetectUtility.detectOutChannel(user, channel)) {
+        if (!detectOutChannel(user, channel)) {
             ErrorMessageUtility.printErrorMessage();
             return;
         }
         jcfUserRepository.outChannel(user, channel);
+    }
+
+    public static boolean detectJoinChannel(User user, Channel channel) {
+        boolean detected = DetectUtility.detect(user) && DetectUtility.detect(channel);
+        boolean notJoined = !user.getChannels().contains(channel);
+        return detected && notJoined;
+    }
+
+    public static boolean detectOutChannel(User user, Channel channel) {
+        boolean detected = DetectUtility.detect(user) && DetectUtility.detect(channel);
+        boolean joined = user.getChannels().contains(channel);
+        return detected && joined;
     }
 }
